+++ Authoring/Unity.Tiny.Rendering.Authoring/AutoMovingDirectionalLight.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using UnityEngine;
 using Unity.Entities;
+using Unity.Entities.Runtime.Build;
 
 namespace Unity.Tiny.Authoring
 {
@@ -12,6 +13,9 @@ namespace Unity.Tiny.Authoring
 
         public void Convert(Entity entity, EntityManager dstManager, GameObjectConversionSystem conversionSystem)
         {
+            if (!conversionSystem.TryGetBuildSettingsComponent<DotsRuntimeBuildProfile>(out _))
+                return;
+
             if (mainCamera == null)
                 throw new ArgumentException($"No camera found in the AutoMovingDirectionalLight authoring component of the gameobject: {name}. Please assign one");