+++ Authoring/Unity.Tiny.Rendering.Authoring/LightingRefProxy.cs
@@ -1,8 +1,0 @@
-using Unity.Entities;
-using Unity.Tiny.Rendering;
-
-[UnityEngine.AddComponentMenu("Hidden/DontUse")]
-[System.Serializable]
-public class LightingRefProxy : SharedComponentDataProxy<LightingRef>
-{
-}