+++ Authoring/Unity.Tiny.Rendering.Authoring/RenderToPassesProxy.cs
@@ -1,8 +1,0 @@
-using Unity.Entities;
-using Unity.Tiny.Rendering;
-
-[UnityEngine.AddComponentMenu("Hidden/DontUse")]
-[System.Serializable]
-public class RenderToPassesProxy : SharedComponentDataProxy<RenderToPasses>
-{
-}